+++ src/main/java/com/example/demo/repository/money/ExpenseMoneyRepository.java
@@ -1,6 +1,7 @@
 package com.example.demo.repository.money;
 
 import com.example.demo.model.money.ExpenseMoney;
+import java.math.BigDecimal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
@@ -12,5 +13,5 @@ public interface ExpenseMoneyRepository
     extends JpaRepository<ExpenseMoney, String>, JpaSpecificationExecutor<ExpenseMoney> {
 
   @Query("SELECT COALESCE(SUM(e.amount), 0) FROM ExpenseMoney e WHERE e.job.id = :jobId")
-  Integer sumByJobId(@Param("jobId") String jobId);
+  BigDecimal sumByJobId(@Param("jobId") String jobId);
 }
